+++ sample/sample5-subscribers/app.ts
@@ -52,15 +52,15 @@ createConnection(options).then(connection => {
                 .createQueryBuilder("p")
                 .leftJoinAndSelect("p.author", "author")
                 .leftJoinAndSelect("p.categories", "categories")
-                .where("p.id = :id", { id: loadedPost.id })
+                .where("p.id = :id", { id: loadedPost!.id })
                 .getOne();
         })
         .then(loadedPost => {
             console.log("---------------------------");
             console.log("load finished. Now lets update entity");
-            loadedPost.text = "post updated";
-            loadedPost.author.name = "Bakha";
-            return postRepository.persist(loadedPost);
+            loadedPost!.text = "post updated";
+            loadedPost!.author.name = "Bakha";
+            return postRepository.persist(loadedPost!);
         })
         .then(loadedPost => {
             console.log("---------------------------");