+++ src/common/DeepPartial.ts
@@ -1,3 +1,6 @@
+/**
+ * Same as Partial<T> but goes deeper and makes Partial<T> all its properties and sub-properties.
+ */
 export type DeepPartial<T> = {
-    readonly [P in keyof T]?: DeepPartial<T[P]>;
+    [P in keyof T]?: DeepPartial<T[P]>;
 };