+++ src/entity-manager/EntityManager.ts
@@ -6,6 +6,9 @@ import {QueryRunnerProviderAlreadyReleasedError} from "../query-runner/error/Que
 import {QueryRunnerProvider} from "../query-runner/QueryRunnerProvider";
 import {ObjectLiteral} from "../common/ObjectLiteral";
 import {FindOneOptions} from "../find-options/FindOneOptions";
+import {DeepPartial} from "../common/DeepPartial";
+import {RemoveOptions} from "../repository/RemoveOptions";
+import {PersistOptions} from "../repository/PersistOptions";
 
 /**
  * Entity manager supposed to work with any entity, automatically find its repository and call its methods,
@@ -29,58 +32,85 @@ export class EntityManager extends BaseEntityManager {
      * Persists (saves) all given entities in the database.
      * If entities do not exist in the database then inserts, otherwise updates.
      */
-    persist<Entity>(entity: Entity): Promise<Entity>;
+    persist<Entity>(entity: Entity, options?: PersistOptions): Promise<Entity>;
 
     /**
      * Persists (saves) all given entities in the database.
      * If entities do not exist in the database then inserts, otherwise updates.
      */
-    persist<Entity>(targetOrEntity: Function, entity: Entity): Promise<Entity>;
+    persist<Entity>(targetOrEntity: Function, entity: Entity, options?: PersistOptions): Promise<Entity>;
 
     /**
      * Persists (saves) all given entities in the database.
      * If entities do not exist in the database then inserts, otherwise updates.
      */
-    persist<Entity>(targetOrEntity: string, entity: Entity): Promise<Entity>;
+    persist<Entity>(targetOrEntity: string, entity: Entity, options?: PersistOptions): Promise<Entity>;
 
     /**
      * Persists (saves) all given entities in the database.
      * If entities do not exist in the database then inserts, otherwise updates.
      */
-    persist<Entity>(entities: Entity[]): Promise<Entity[]>;
+    persist<Entity>(entities: Entity[], options?: PersistOptions): Promise<Entity[]>;
 
     /**
      * Persists (saves) all given entities in the database.
      * If entities do not exist in the database then inserts, otherwise updates.
      */
-    persist<Entity>(targetOrEntity: Function, entities: Entity[]): Promise<Entity[]>;
+    persist<Entity>(targetOrEntity: Function, entities: Entity[], options?: PersistOptions): Promise<Entity[]>;
 
     /**
      * Persists (saves) all given entities in the database.
      * If entities do not exist in the database then inserts, otherwise updates.
      */
-    persist<Entity>(targetOrEntity: string, entities: Entity[]): Promise<Entity[]>;
+    persist<Entity>(targetOrEntity: string, entities: Entity[], options?: PersistOptions): Promise<Entity[]>;
 
     /**
      * Persists (saves) a given entity in the database.
      */
-    persist<Entity>(targetOrEntity: (Entity|Entity[])|Function|string, maybeEntity?: Entity|Entity[]): Promise<Entity|Entity[]> {
+    persist<Entity>(targetOrEntity: (Entity|Entity[])|Function|string, maybeEntity?: Entity|Entity[], options?: PersistOptions): Promise<Entity|Entity[]> {
         const target = arguments.length === 2 ? maybeEntity as Entity|Entity[] : targetOrEntity as Function|string;
         const entity = arguments.length === 2 ? maybeEntity as Entity|Entity[] : targetOrEntity as Entity|Entity[];
         return Promise.resolve().then(() => { // we MUST call "fake" resolve here to make sure all properties of lazily loaded properties are resolved.
             if (typeof target === "string") {
-                return this.getRepository<Entity|Entity[]>(target).persist(entity);
+                return this.getRepository<Entity|Entity[]>(target).persist(entity, options);
             } else {
+                // todo: throw exception if constructor in target is not set
                 if (target instanceof Array) {
                     if (target.length === 0)
                         return Promise.resolve(target);
 
-                    return this.getRepository<Entity[]>(target[0].constructor).persist(entity as Entity[]);
+                    return this.getRepository<Entity[]>(target[0].constructor).persist(entity as Entity[], options);
                 } else {
-                    return this.getRepository<Entity>(target.constructor).persist(entity as Entity);
+                    return this.getRepository<Entity>(target.constructor).persist(entity as Entity, options);
                 }
             }
         });
+    }
+
+    /**
+     * Updates entity partially. Entity can be found by a given conditions.
+     */
+    async update<Entity>(target: Function|string, conditions: Partial<Entity>, partialEntity: DeepPartial<Entity>, options?: PersistOptions): Promise<void>;
+
+    /**
+     * Updates entity partially. Entity can be found by a given find options.
+     */
+    async update<Entity>(target: Function|string, findOptions: FindOneOptions<Entity>, partialEntity: DeepPartial<Entity>, options?: PersistOptions): Promise<void>;
+
+    /**
+     * Updates entity partially. Entity can be found by a given conditions.
+     */
+    async update<Entity>(target: Function|string, conditionsOrFindOptions: Partial<Entity>|FindOneOptions<Entity>, partialEntity: DeepPartial<Entity>, options?: PersistOptions): Promise<void> {
+        return this.getRepository<Entity|Entity[]>(target as any)
+            .update(conditionsOrFindOptions as any, partialEntity, options);
+    }
+
+    /**
+     * Updates entity partially. Entity will be found by a given id.
+     */
+    async updateById<Entity>(target: Function|string, id: any, partialEntity: DeepPartial<Entity>, options?: PersistOptions): Promise<void> {
+        return this.getRepository<Entity|Entity[]>(target as any)
+            .updateById(id, partialEntity, options);
     }
 
     /**
@@ -91,43 +121,51 @@ export class EntityManager extends BaseEntityManager {
     /**
      * Removes a given entity from the database.
      */
-    remove<Entity>(targetOrEntity: Function, entity: Entity): Promise<Entity>;
+    remove<Entity>(targetOrEntity: Function, entity: Entity, options?: RemoveOptions): Promise<Entity>;
 
     /**
      * Removes a given entity from the database.
      */
-    remove<Entity>(targetOrEntity: string, entity: Entity): Promise<Entity>;
+    remove<Entity>(targetOrEntity: string, entity: Entity, options?: RemoveOptions): Promise<Entity>;
 
     /**
      * Removes a given entity from the database.
      */
-    remove<Entity>(entity: Entity[]): Promise<Entity>;
+    remove<Entity>(entity: Entity[], options?: RemoveOptions): Promise<Entity>;
 
     /**
      * Removes a given entity from the database.
      */
-    remove<Entity>(targetOrEntity: Function, entity: Entity[]): Promise<Entity[]>;
+    remove<Entity>(targetOrEntity: Function, entity: Entity[], options?: RemoveOptions): Promise<Entity[]>;
 
     /**
      * Removes a given entity from the database.
      */
-    remove<Entity>(targetOrEntity: string, entity: Entity[]): Promise<Entity[]>;
+    remove<Entity>(targetOrEntity: string, entity: Entity[], options?: RemoveOptions): Promise<Entity[]>;
 
     /**
      * Removes a given entity from the database.
      */
-    remove<Entity>(targetOrEntity: (Entity|Entity[])|Function|string, maybeEntity?: Entity|Entity[]): Promise<Entity|Entity[]> {
+    remove<Entity>(targetOrEntity: (Entity|Entity[])|Function|string, maybeEntity?: Entity|Entity[], options?: RemoveOptions): Promise<Entity|Entity[]> {
         const target = arguments.length === 2 ? maybeEntity as Entity|Entity[] : targetOrEntity as Function|string;
         const entity = arguments.length === 2 ? maybeEntity as Entity|Entity[] : targetOrEntity as Entity|Entity[];
         if (typeof target === "string") {
-            return this.getRepository<Entity|Entity[]>(target).remove(entity);
+            return this.getRepository<Entity|Entity[]>(target).remove(entity, options);
         } else {
+            // todo: throw exception if constructor in target is not set
             if (target instanceof Array) {
-                return this.getRepository<Entity[]>(target[0].constructor).remove(entity as Entity[]);
+                return this.getRepository<Entity[]>(target[0].constructor).remove(entity as Entity[], options);
             } else {
-                return this.getRepository<Entity>(target.constructor).remove(entity as Entity);
+                return this.getRepository<Entity>(target.constructor).remove(entity as Entity, options);
             }
         }
+    }
+
+    /**
+     * Removes entity by a given entity id.
+     */
+    async removeById(targetOrEntity: Function|string, id: any, options?: RemoveOptions): Promise<void> {
+        return this.getRepository(targetOrEntity as any).removeById(id, options);
     }
 
     /**