+++ test/functional/persistence/partial-persist/entity/Counters.ts
@@ -1,0 +1,16 @@
+import {EmbeddableEntity} from "../../../../../src/decorator/entity/EmbeddableEntity";
+import {Column} from "../../../../../src/decorator/columns/Column";
+
+@EmbeddableEntity()
+export class Counters {
+
+    @Column()
+    stars: number;
+
+    @Column()
+    commentCount: number;
+
+    @Column()
+    metadata: string;
+
+}