+++ test/integration/sample3-many-to-one.ts
@@ -65,7 +65,7 @@ describe("many-to-one", function() {
             newPost.text = "Hello post";
             newPost.title = "this is post title";
             newPost.details = details;
-            return postRepository.persist(newPost).then(post => savedPost = post);
+            return postRepository.persist(newPost).then(post => savedPost = post as Post);
         });
 
         it("should return the same post instance after its created", function () {
@@ -188,7 +188,7 @@ describe("many-to-one", function() {
             newPost.title = "this is post title";
             newPost.category = category;
 
-            return postRepository.persist(newPost).then(post => savedPost = post);
+            return postRepository.persist(newPost).then(post => savedPost = post as Post);
         });
 
         it("should return the same post instance after its created", function () {
@@ -267,13 +267,13 @@ describe("many-to-one", function() {
 
             return postRepository
                 .persist(newPost)
-                .then(post => savedPost = post);
+                .then(post => savedPost = post as Post);
         });
 
         it("should ignore updates in the model and do not update the db when entity is updated", function () {
             newPost.details.comment = "i am updated comment";
             return postRepository.persist(newPost).then(updatedPost => {
-                updatedPost.details.comment.should.be.equal("i am updated comment");
+                updatedPost.details!.comment!.should.be.equal("i am updated comment");
                 return postRepository
                     .createQueryBuilder("post")
                     .leftJoinAndSelect("post.details", "details")
@@ -281,7 +281,7 @@ describe("many-to-one", function() {
                     .setParameter("id", updatedPost.id)
                     .getOne();
             }).then(updatedPostReloaded => {
-                updatedPostReloaded.details.comment.should.be.equal("this is post");
+                updatedPostReloaded!.details.comment.should.be.equal("this is post");
             });
         }); // todo: also check that updates throw exception in strict cascades mode
     });
@@ -305,7 +305,7 @@ describe("many-to-one", function() {
 
             return postRepository
                 .persist(newPost)
-                .then(post => savedPost = post);
+                .then(post => savedPost = post as Post);
         });
 
         it("should ignore updates in the model and do not update the db when entity is updated", function () {
@@ -318,7 +318,7 @@ describe("many-to-one", function() {
                     .setParameter("id", updatedPost.id)
                     .getOne();
             }).then(updatedPostReloaded => {
-                updatedPostReloaded.details.comment.should.be.equal("this is post");
+                updatedPostReloaded!.details.comment.should.be.equal("this is post");
             });
         });
     });
@@ -340,12 +340,12 @@ describe("many-to-one", function() {
             return postImageRepository
                 .persist(newImage)
                 .then(image => {
-                    savedImage = image;
-                    newPost.image = image;
+                    savedImage = image as PostImage;
+                    newPost.image = image as PostImage;
                     return postRepository.persist(newPost);
 
                 }).then(post => {
-                    newPost = post;
+                    newPost = post as Post;
                     return postRepository
                         .createQueryBuilder("post")
                         .leftJoinAndSelect("post.image", "image")
@@ -354,8 +354,8 @@ describe("many-to-one", function() {
                         .getOne();
 
                 }).then(loadedPost => {
-                    loadedPost.image.url = "new-logo.png";
-                    return postRepository.persist(loadedPost);
+                    loadedPost!.image.url = "new-logo.png";
+                    return postRepository.persist(loadedPost!);
 
                 }).then(() => {
                     return postRepository
@@ -366,7 +366,7 @@ describe("many-to-one", function() {
                         .getOne();
                     
                 }).then(reloadedPost => {
-                    reloadedPost.image.url.should.be.equal("new-logo.png");
+                    reloadedPost!.image.url.should.be.equal("new-logo.png");
                 });
         });
 
@@ -389,12 +389,12 @@ describe("many-to-one", function() {
             return postMetadataRepository
                 .persist(newMetadata)
                 .then(metadata => {
-                    savedMetadata = metadata;
-                    newPost.metadata = metadata;
+                    savedMetadata = metadata as PostMetadata;
+                    newPost.metadata = metadata as PostMetadata;
                     return postRepository.persist(newPost);
 
                 }).then(post => {
-                    newPost = post;
+                    newPost = post as Post;
                     return postRepository
                         .createQueryBuilder("post")
                         .leftJoinAndSelect("post.metadata", "metadata")
@@ -403,8 +403,8 @@ describe("many-to-one", function() {
                         .getOne();
 
                 }).then(loadedPost => {
-                    loadedPost.metadata = null;
-                    return postRepository.persist(loadedPost);
+                    loadedPost!.metadata = null;
+                    return postRepository.persist(loadedPost!);
 
                 }).then(() => {
                     return postRepository
@@ -415,7 +415,7 @@ describe("many-to-one", function() {
                         .getOne();
 
                 }).then(reloadedPost => {
-                    expect(reloadedPost.metadata).to.be.empty;
+                    expect(reloadedPost!.metadata).to.be.empty;
                 });
         });
 
@@ -436,7 +436,7 @@ describe("many-to-one", function() {
             details.posts = [];
             details.posts.push(newPost);
 
-            return postDetailsRepository.persist(details).then(details => savedDetails = details);
+            return postDetailsRepository.persist(details).then(details => savedDetails = details as PostDetails);
         });
 
         it("should return the same post instance after its created", function () {